Add sign-in page design
